LayoutTests/fast/page-color-sampling/color-sampling-horizontal-overflow-with-viewport-meta.html

<!DOCTYPE html> <!-- webkit-test-runner [ ContentInsetBackgroundFillEnabled=true pageTopColorSamplingEnabled=true useFlexibleViewport=true obscuredInset.top=50 ] -->
<html>
<meta name="viewport" content="width=device-width">
<head>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/ui-helper.js"></script>
<style>
body, html {
    font-family: system-ui;
    font-size: 16px;
    margin: 0;
    padding: 0;
    background: rgb(42, 42, 42);
    width: 100%;
    height: 100%;
}

.fixed-header {
    position: fixed;
    top: 0;
    left: 0;
    width: 100vw;
    height: 60px;
    background: rgb(0, 122, 255);
    z-index: 100;
}

.wide {
    width: 640px;
    height: 100px;
    background: red;
    margin-top: 50vw;
}

.tall {
    height: 1000px;
}
</style>
<script src="../../resources/js-test.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script>
jsTestIsAsync = true;

addEventListener("load", async () => {
    await UIHelper.setObscuredInsets(50, 0, 0, 0);
    await UIHelper.ensurePresentationUpdate();

    color = (await UIHelper.fixedContainerEdgeColors()).top;
    shouldBeEqualToString("color", "rgb(0, 122, 255)");

    finishJSTest();
});
</script>
</head>
<body>
    <div class="fixed-header"></div>
    <div class="tall"></div>
    <div class="wide"></div>
    <div class="tall"></div>
</body>
</html>
