LayoutTests/fast/page-color-sampling/color-sampling-horizontal-overflow-with-viewport-meta-expected.txt

PASS color is "rgb(0, 122, 255)"
PASS successfullyParsed is true

TEST COMPLETE
